better error handling
The Logger.error call returns an :ok atom, but the downstream
BroadwaySQS.Producer.handle_receive_messages/1 expects a list
it can call `length/1` on

The net result is that if bad credentials are given in the
config of a module that uses Broadway, then the GenServer keeps
running instead of terminating just because one of the calls to
AWS gives an error. It may be temporary. Returning an empty list
seems reasonable