Crash on ack error
Before this patch we were silently ignoring ack failures.

We use `ExAws.request` (non-bang version) in only one other place,
https://github.com/dashbitco/broadway_sqs/blob/v0.5.0/lib/broadway_sqs/ex_aws_client.ex#L50,
however we're correctly unwrapping result and logging an error. We could
change that callsite to use `request!` too, but I wasn't sure if it was
important to gracefully handle that failure and return `[]`, see:
https://github.com/dashbitco/broadway_sqs/blob/v0.5.0/test/broadway_sqs/ex_aws_client_test.exs#L342:L343